[MERGE]: Merged hamza-qorb-new into main

+++ core/app/Http/Controllers/Dashboard/EmployeeController.php
@@ -1,6 +1,8 @@
 <?php
 
 namespace App\Http\Controllers\Dashboard;
+
+use App\Helpers\Helper;
 use App\Http\Controllers\Controller;
 use App\Models\Employee;
 use App\Models\Role;
@@ -24,6 +26,9 @@ class EmployeeController extends Controller
      */
     public function index()
     {
+        if (!Helper::checkPermission(8)) {
+            return redirect()->route('NoPermission');
+        }
         // Check Permissions
         if (!@Auth::user()->permissionsGroup->settings_status) {
             return redirect()->route('adminHome');
@@ -47,6 +52,9 @@ class EmployeeController extends Controller
      */
     public function create()
     {
+        if (!Helper::checkPermission(8)) {
+            return redirect()->route('NoPermission');
+        }
         // Check Permissions
         if (!@Auth::user()->permissionsGroup->settings_status) {
             return redirect()->route('NoPermission');
@@ -66,6 +74,9 @@ class EmployeeController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Helper::checkPermission(8)) {
+            return redirect()->route('NoPermission');
+        }
         // Check Permissions
         if (!@Auth::user()->permissionsGroup->settings_status) {
             return redirect()->route('NoPermission');
@@ -126,6 +137,9 @@ class EmployeeController extends Controller
      */
     public function edit($id)
     {
+        if (!Helper::checkPermission(8)) {
+            return redirect()->route('NoPermission');
+        }
         // Check Permissions
         if (!@Auth::user()->permissionsGroup->settings_status && @Auth::user()->id != $id) {
             return redirect()->route('NoPermission');
@@ -133,15 +147,11 @@ class EmployeeController extends Controller
         // General for all pages
         $GeneralWebmasterSections = WebmasterSection::where('status', '=', '1')->orderby('row_no', 'asc')->get();
         // General END
-        $roles = Role::orderby('id', 'asc')->get();
 
-        if (@Auth::user()->permissionsGroup->view_status) {
-            $employee = Employee::where('created_by', '=', Auth::user()->id)->orwhere('id', '=', Auth::user()->id)->find($id);
-        } else {
-            $employee = Employee::find($id);
-        }
+        $employee = User::find($id);
+
         if (!empty($employee)) {
-            return view("dashboard.employees.edit", compact("employee", "roles", "GeneralWebmasterSections"));
+            return view("dashboard.employees.edit", compact("employee", "GeneralWebmasterSections"));
         } else {
             return redirect()->action('Dashboard\EmployeeController@index');
         }
@@ -152,11 +162,14 @@ class EmployeeController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!Helper::checkPermission(8)) {
+            return redirect()->route('NoPermission');
+        }
         // Check Permissions
         if (!@Auth::user()->permissionsGroup->settings_status && @Auth::user()->id != $id) {
             return redirect()->route('NoPermission');
         }
-        $employee = Employee::find($id);
+        $employee = User::find($id);
 
         $this->validate($request, [
             'name' => 'required',
@@ -164,7 +177,7 @@ class EmployeeController extends Controller
             'phone' => 'required|string|regex:/^([0-9\s\-\+\(\)]*)$/|unique:employees,phone,'. $employee->id,
             'password' => 'required|min:6',
             'photo' => 'mimes:png,jpeg,jpg,gif,svg',
-            'role_id' => 'required'
+            'role' => 'required'
         ]);
 
         if (!empty($employee)) {
@@ -195,7 +208,7 @@ class EmployeeController extends Controller
                 $employee->email = $request->email;
                 $employee->phone = $request->phone;
                 $employee->password = bcrypt($request->password);
-                $employee->role_id = $request->role_id;
+                $employee->role = $request->role;
                 
                 if ($request->photo_delete == 1) {
                     // Delete a User file
@@ -228,11 +241,14 @@ class EmployeeController extends Controller
      */
     public function destroy($id)
     {
+        if (!Helper::checkPermission(8)) {
+            return redirect()->route('NoPermission');
+        }
         // Check Permissions
         if (!@Auth::user()->permissionsGroup->settings_status) {
             return redirect()->route('NoPermission');
         }
-        $employee = Employee::find($id);
+        $employee = User::find($id);
             if ($employee->photo != "") {
                 File::delete($this->getUploadPath() . $employee->photo);
             }
